Submit level gate keeper two

+++ src/Denial/Denial.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.10;
 
-import 'openzeppelin-contracts/contracts/utils/math/SafeMath.sol';
+import "openzeppelin-contracts/contracts/utils/math/SafeMath.sol";
 
 contract Denial {
-
     using SafeMath for uint256;
+
     address public partner; // withdrawal partner - pay the gas, split the withdraw
     // Foundry: wrapped address in "payable" to avoid error
     address payable public constant owner = payable(address(0xA9E));
-    uint timeLastWithdrawn;
-    mapping(address => uint) withdrawPartnerBalances; // keep track of partners balances
+    uint256 timeLastWithdrawn;
+    mapping(address => uint256) withdrawPartnerBalances; // keep track of partners balances
 
     function setWithdrawPartner(address _partner) public {
         partner = _partner;
@@ -18,10 +18,10 @@ contract Denial {
 
     // withdraw 1% to recipient and 1% to owner
     function withdraw() public {
-        uint amountToSend = address(this).balance.div(100);
+        uint256 amountToSend = address(this).balance.div(100);
         // perform a call without checking return
         // The recipient can revert, the owner will still get their share
-        partner.call{value:amountToSend}("");
+        partner.call{ value: amountToSend }("");
         owner.transfer(amountToSend);
         // keep track of last withdrawal time
         // Foundry: changed from "now" as it is deprecated
@@ -30,10 +30,10 @@ contract Denial {
     }
 
     // allow deposit of funds
-    receive() external payable {}
+    receive() external payable { }
 
     // convenience function
-    function contractBalance() public view returns (uint) {
+    function contractBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
