+++ src/DexTwo/DexTwoHack.sol
@@ -1,19 +1,21 @@
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.10;
+
 import "./DexTwo.sol";
 
 contract DexTwoHack {
     DexTwo dexTwo;
+
     constructor(DexTwo _dexTwo) {
         dexTwo = _dexTwo;
     }
 
-    function balanceOf(address) public view returns(uint balance){
+    function balanceOf(address) public view returns (uint256 balance) {
         balance = 1;
     }
 
-    function transferFrom(address from, address to, uint amount) public returns(bool) {
+    function transferFrom(address from, address to, uint256 amount) public returns (bool) {
         return true;
     }
